fix: use full GUID entropy for advisory lock keys
XORs both 64-bit halves of the GUID to preserve all 16 bytes.
Narrows the dual-key overload parameters from long to int.
Removes lossy truncating casts in the SQL interpolation.

+++ ClinicFlow.Domain/Interfaces/IUnitOfWork.cs
@@ -15,7 +15,7 @@ public interface IUnitOfWork
 
     Task<TResult> ExecuteWithLockAsync<TResult>(
         Guid lockKeyPart1,
-        long lockKeyPart2,
+        int lockKeyPart2,
         Func<CancellationToken, Task<TResult>> operation,
         CancellationToken cancellationToken = default
     );
